user_service/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Zia Khan <zia@panacloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = {extras = ["standard"], version = "^0.27.1"}
sqlmodel = "^0.0.16"
psycopg = {extras = ["binary"], version = "^3.1.18"}
fastapi = "^0.111.0"
psycopg2-binary = "^2.9.9"
protobuf = "^5.27.1"
aiokafka = "^0.10.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
bcrypt = "^4.2.0"
